Add array uri to tiledb_array_deserialize (#4961)
[sc-47129

](https://app.shortcut.com/tiledb-inc/story/47129/add-array-uri-to-tiledb-array-deserialize
)
Today `tiledb_array_deserialize` is using a hardcoded uri name
(`"deserialized_array"`) , whereas `tiledb_deserialize_query_and_array`
is requiring the actual `array_uri` as an argument. The former is
problematic for wrapping those calls in python, so this ticket is aiming
to unify the two to both receive the `array_uri` as input. (history on
why this is needed here:
https://app.shortcut.com/tiledb-inc/story/47139/investigate-if-core-serialization-apis-could-be-used-externally-to-pass-cache-open-arrays)

There's no backward compatibility issue here as this is a
deserialization function only called by REST Server, so we don't need to
keep both versions of `tiledb_array_deserialize`.

That said, the REST-CI tests that run based on latest dev on Core and
latest master in Cloud-REST will break after this change. We will handle
this by adapting the REST-CI script to patch temporarily Cloud-REST
master with the needed change and remove when 2.24 is released
https://github.com/TileDB-Inc/TileDB-Internal/pull/27 . Another idea was
to maintain both flavors until the release is ready to go out and then
remove but C doesn't support overloading functions.

Future work to track:  https://github.com/TileDB-Inc/TileDB/pull/4961

---
TYPE: IMPROVEMENT
DESC: Add array uri to tiledb_array_deserialize