Overhaul `tdb_gzip_embedded_data`.
It was rewritten from C++ to CMake, and compression is now being done with CMake's commands.
This allows us to pack the upstream `magic.mgc` file and stop keeping a pre-compressed and pre-escaped one in source control.
The size of the uncompressed file used to be kept at the start of the binary file. We no longer have the capability to easily modify binary files with CMake, so the script generates a complete header, alongside a constexpr variable with the uncompressed size.