Fix `DetectStdPmr` for cross-compiling scenarios. (#4995)
[SC-47807](https://app.shortcut.com/tiledb-inc/story/47807/detectstdpmr-fails-in-macos-when-cross-compiling)

We use `try_compile` instead of `try_run` to detect whether the headers
for `std::pmr` are available. This fixes errors in cross-compiling
scenarios, like when building for osx-arm64 in an osx-x64 machine.
Should fix failures in conda-forge/TileDB-feedstock#290, waiting for
adding a patch and validating.

---
TYPE: BUILD
DESC: Fix configuration errors when cross-compiling.