Implement consolidation memory budget variables.
This implement 3 new variables for memory budgetting for consolidation. The values are a set of 3 weights that weigh each parts of the consolidation algorithm. The buffers, the read operation and the write operation. All weights are summed and ratios are computed to divide sm.mem.total_budget. The values are:
* `sm.mem.consolidation.buffers_weight`
* `sm.mem.consolidation.reader_weight`
* `sm.mem.consolidation.writer_weight`

---
TYPE: IMPROVEMENT
DESC: Implement consolidation memory budget variables.