Fix SEG in the SparseGlobalOrderReader when using a Query Condition. (#5012)
The lack of this check was causing a SEG fault when reading with a query
condition with the Sparse Global Order reader. More specifically, the
code previously was trying to access tiles deleted from memory in some
cases. These tiles were deleted as they were marked as ```unused```
during the read process.

[sc-47713]

---
TYPE: BUG
DESC: Fix SEG faults in the SparseGlobalOrderReader when using a Query
Condition.

---------

Co-authored-by: KiterLuc <67824247+KiterLuc@users.noreply.github.com>