Fix output datatype for some filters. (#4988)
Story details: https://app.shortcut.com/tiledb-inc/story/47328

The addition of `Filter::output_datatype` to validate filter pipelines
prevented some invalid pipelines from being constructed; but it depends
on the correctness of each `output_datatype` implementation to ensure
that a filter can run correctly over the output of a previous filter.

For the Rust bindings we have built randomized test strategies which
assemble pipelines over combinations of filters which are unlikely to
have appeared in the wild. This has exposed some missing or incorrect
`output_datatype` implementations which allow filters which depend on
certain input properties to appear in pipelines erroneously.

This pull request demonstrates some such cases in unit tests using the
XOR and Delta compression filters, which depend on a particular bit
width of their input, and fixes the `output_datatype` implementations of
filters which do not preserve the datatype between input and output.

---
TYPE: BUG
DESC: Fix output datatype for some filters.