Improve detection of macOS deployment target, for choosing pmr implementation. (#5027)
SC-48427

#4995 changed detection of `std::pmr` from trying to _run_ a C++ file
referencing pmr APIs, to trying to _compile_ it, which works in cross
compilation scenarios and fixed Conda failures for reasons I do not
understand.

However in macOS versions prior to 14 the pmr headers exist, but
binaries compiled with it fail ot run. There was already a check to
force using the vendored pmr in such versions but it was defective,
because it checked the `MACOS_DEPLOYMENT_TARGET` environment variable,
which might not exist. Before #4995, the runtime failure on these
versions would be caught by `try_run`. This PR updates the detection
script to use the
[`CMAKE_OSX_DEPLOYMENT_TARGET`](https://cmake.org/cmake/help/latest/variable/CMAKE_OSX_DEPLOYMENT_TARGET.html)
variable, which gets automatically filled by CMake either from the
aforementioned environment variable or automatically.

---
TYPE: NO_HISTORY