Add simplified C++23 `chunk_view` (#5035)
This PR adds a class to provide a subset of the functionality of the
C++23 `std::ranges::chunk_view`. This view divides an underlying view
into a range of equal-sized chunks (the last one may be shorter than the
others). A complete description of `chunk_view` can be found on
cppreference.com, which also includes "exposition only" aspects that
point to particular implementation approaches. We did not adopt those
approaches in most cases, though the public API of the class matches the
spec, for the functions that we did implement. This view will be used to
partition the original input data from the user query into, well, chunks
that will be individually sorted and then written to a TileDB array as a
data block. Data blocks are of fixed and uniform byte size that will
contain a chunk smaller than or equal to its byte limit.

* The class assumes a` random_access_range`, which is what we have in
our buffers that we are chunking, so I didn't try to generalize any
further than that.

* The class provides a complete iterator interface via `iterator_facade`
as well as a complete view interface via `view_interface`. Aspects of
those interfaces that rely on C++23 are pended.

* Unit tests validate that `chunk_view` and its iterator meet the
expected `std::ranges` concepts.

* The unit tests do not yet test composition of this view with other
views (our custom views as well as other C++20 views). This will be
fairly extensive and the subject of its own PR.

* One todo would be to create a variable chunk class view that chunks up
the underlying view into variable-sized pieces, the better to fit into
the data blocks. However, the data blocks will be fairly large, so there
should not be too much internal fragmentation. We can always manage
internal fragmentation by making the chunks some fraction of the block
size, so even if there is some variation in chunk size, they can be
well-packed into the data blocks.

---
TYPE: IMPROVEMENT
DESC: Implementation of a `chunk_view` class to provide a subset of
C++23 chunk_view, suitable for supporting external sort.

---------

Co-authored-by: Luc Rancourt <lucrancourt@gmail.com>