Add temporary mitigations for CI failures in `windows-2022`. (#5057)
[SC-48975](https://app.shortcut.com/tiledb-inc/story/48975)

Visual Studio 2022 17.10 has been recently released and one of its
changes to MSVC is that the constructor of `std::mutex` became
`constexpr`. According to MSVC's binary compatibility guarantees, a
program must use a Visual C++ redistributable version at least as new as
the toolset that compiled the program. However, due to [a
defect](https://github.com/actions/runner-images/issues/10004) on the
GitHub Actions `windows-2022` image, an old redistributable version was
picked, causing weird failures in CI.

This PR fixes these failures, by defining the
`_DISABLE_CONSTEXPR_MUTEX_CONSTRUCTOR` symbol, which reverts the change
to `std::mutex`. Due to unexplained errors, I could not do that for the
standalone unit test workflow, and instead I changed that to use the
`windows-2019` image which does not exhibit this problem.

This PR should be reverted once the issue on GHA gets fixed.

---
TYPE: NO_HISTORY