Tests to exercise combinations of views for external sort. (#5051)
This PR provides the file `tiledb/common/util/test/unit_view_combo.cc`
that includes a set of tests of various combinations of the view classes
that have been developed for external sort. In particular, we have tests
to `chunk` a `chunk_view`, `var_length_view`, and `alt_var_length_view`,
to `chunk` a `permutation_view` and permute a `chunk_view`, and to
`chunk` a `zip_view` and `zip` a `chunk_view`.

The tests effect these compositions by defining a range, taking a view
of it (by calling, e.g., `chunk()`), and then calling another view
adaptor. We do not intend to support composition by `operator|()` at
this time.

[sc-48753]

---
TYPE: NO_HISTORY
DESC: Tests to exercise combinations of views.

---------

Co-authored-by: Luc Rancourt <lucrancourt@gmail.com>
Co-authored-by: KiterLuc <67824247+KiterLuc@users.noreply.github.com>