Add the directory to the tiledb library on PATH in `tiledb_regression`. (#5088)
After https://github.com/TileDB-Inc/TileDB/pull/5054,
`tiledb_regression` started to run in the nightlies. `tiledb_regression`
directly links to the `tiledb` target, which on Windows when building a
shared library, is not found by default. This PR updates the
`tiledb_regression` CTest test to append the path to `tiledb.dll` to
PATH so that we don't have issues running the nightlies anymore.

[SC-48054]

---
TYPE: NO_HISTORY