Add design doc diagram (#5086)
This PR adds the active diagram associated with the external sort design
documentation. It is in SVG format to allow portable editing with
multiple tools, as well as to be more friendly to GitHub. It has been
verified to open and be editable in InkScape, which is freely available
in MacOS, Linux, and Windows. For compatibility, it is recommended that
further editing of the diagram be done exclusively in InkScape.

---
TYPE: NO_HISTORY
DESC: Add SVG format design documentation diagram