Disable nightly testing of HDFS and log a warning when HDFS is enabled. (#5085)
[SC-49481](https://app.shortcut.com/tiledb-inc/story/49481/disable-testing-hdfs-in-nightlies)

[SC-49482](https://app.shortcut.com/tiledb-inc/story/49482/add-deprecation-warning-when-configuring-with-hdfs-enabled)

This PR disables testing HDFS in nightlies. The runs have been very
brittle for years when more tests were added in and with the recent
addition of many tests to all VFSes, it is now impossible to get a
successful run. As we are planning to fully remove support for HDFS in
TileDB version 2.28, we have made the decision to not investigate the
failures. The Core will still be built with HDFS enabled to catch
build-time regressions.

Furthermore, the build script was updated to emit a warning when
configuring with `-DTILEDB_HDFS=ON`.

---
TYPE: DEPRECATION
DESC: The HDFS backend is no longer officially tested by TileDB. As
announced before, it is scheduled to be removed in version 2.28, to be
released in Q4 2024.