Current Domain C API implementation. (#5096)
This implement the CAPI for the current domain feature which defines an
area of the domain that is currently in use for the array. The
information lives as part of the schema of the array and can be evolved
to expand the domain using the schema evolution APIs. The current domain
is currently set through an N-dimensional rectangle but support for
different shape types might be supported in the future.

[sc-47337]

---
TYPE: CAPI
DESC: Current Domain C API implementation.

---------

Co-authored-by: KiterLuc <67824247+KiterLuc@users.noreply.github.com>
Co-authored-by: Luc Rancourt <lucrancourt@gmail.com>