Fix ASAN nightlies with failure in unit_query.
After moving more tests to run in the ASAN nightlies, we encountered an issue in unit_query where it doesn't properly setup a test offset buffer with an extra offset and tries to read the extra offset when copying over to the tile.

[sc-49797]

---
TYPE: NO_HISTORY
DESC: Fix ASAN nightlies with failure in unit_query.