Fix ASAN nightlies with failure in unit_cppapi_string.
After moving more tests to run in the ASAN nightlies, we encountered an issue in unit_cppapi_string where a handle is used after being freed. This test is doing this by design so this PR disables the test when ASAN is enabled.

[sc-49797]

---
TYPE: NO_HISTORY
DESC: Fix ASAN nightlies with failure in unit_cppapi_string.