Warn users using dense arrays with sparse fragments.
This adds a warning to user using dense arrays with sparse fragments that support will be removed in TileDB 2.27. Support for writing sparse fragments in dense arrays was removed in TileDB 2.5 and as the user can easily consolidate to get rid of sparse fragments, the decision was made to remove support for reading these arrays.

---
TYPE: DEPRECATION
DESC: Warn users using dense arrays with sparse fragments.