Add memory tracking for RTree ranges. (#5029)
[SC-46801](https://app.shortcut.com/tiledb-inc/story/46801/add-memory-tracking-for-rtree-ranges)

This PR enables memory tracking for RTree ranges. The first part
consists of updating the range class to be backed by a
`tdb::pmr::pmr_vector`[^1] instead of an `std::vector` which allows
users to specify a custom allocator that tracks memory, in an opt-in
basis. The second part is updating the RTree class to use these
allocators when creating the ranges.

Validated locally. Let me know what kind of testing to add.

[^1]: We are _not_ using `tdb::pmr::vector`, because it is not copy or
move constructible, which would break other uses of ranges. Only ranges
created by RTrees will be tracked in this PR.

---
TYPE: NO_HISTORY

---------

Co-authored-by: KiterLuc <67824247+KiterLuc@users.noreply.github.com>
Co-authored-by: Luc Rancourt <lucrancourt@gmail.com>