Perform validity checks for ranges to add in an `NDRectangle`. (#5144)
[SC-49157](https://app.shortcut.com/tiledb-inc/story/49157/check-and-test-out-of-order-ranges-for-currentdomain)

This PR adds validation that the ranges added in an `NDRectangle` are in
the correct order. We add a `check_range_is_valid` function that accepts
a range and a datatype, and call it on `NDRectangle::set_range` with the
dimension's data type. The generic `check_range_is_valid` was also
updated to handle `std::string_view`.

---
TYPE: NO_HISTORY