Fix serialization issue with schema evolution for query v3. (#5154)
Query v3 requires fragment metadata serialization/deserialization inside
the array object. However fragment metadata capnp model wasn't including
the name of the array schema that each fragment was written with , and
in case of schema evolution this was leading to serializing the wrong
schema (always the latest) and making the REST server read the wrong
data from disk.

This PR adds the missing array schema name field and adapts the
ser/deser to pick the schema based on that name.
[sc-48707]

---
TYPE: BUG
DESC: Fix serialization issue with schema evolution for query v3.