Warn users using dense arrays with sparse fragments. (#5116)
Support for reading sparse fragments in dense arrays will be removed in
TileDB version 2.27 to be released in September 2024. To make sure these
arrays continue to work after an upgrade to version 2.27 or later,
please consolidate the sparse fragments using a TileDB version 2.26 or
earlier.

For now, this adds a warning to user using dense arrays with sparse
fragment. Support for writing sparse fragments in dense arrays was
removed in TileDB 2.5 and as the user can easily consolidate to get rid
of sparse fragments, the decision was made to remove support for reading
these arrays.

[sc-47348]

---
TYPE: DEPRECATION
DESC: Warn users using dense arrays with sparse fragments.