`class RestClient` C.41 conformance. (#4994)
The centerpiece of this PR is to bring `class RestClient` into
conformance with C.41. Full initialization of that class was happening
not only in its `init()` member function but also in
`StorageManager::set_default_tags`. The PR removes both of these
functions and moves their behavior to `RestClient::RestClient`, which
has a new signature and much-expanded body.

An interrelated goal is to extricate `class StorageManager` from the
implementation of C API function `tiledb_ctx_set_tag`; the new
implementation calls `RestClient` directly. Not only was initialization
for this happening outside of `RestClient`, but the underlying data
structure was duplicated in `StorageManager` and then only used to run a
few tests. This PR removes the relevant member functions and variables
from `class StorageManager`, reworks the C API function to call
`RestClient` directly, adds `RestClient::extra_headers()` to replace
`StorageManager::tags()`, and reworks the C API tests accordingly.

This PR has a bit of ancillary support work. It adds
`Context::has_rest_client()` and adds a proxy for this and
`Context::rest_client()` in the C API handle class for `class Context`.
It updates use of `Config::get` to non-deprecated varieties.

This PR remove all preprocessor use of the configuration variable
`TILEDB_SERIALIZATION` from `class RestClient`. Instead of using the
preprocessor, there is now a server-disabled base class and a
server-enabled derived class. The server-enabled class is now declared
and defined in separate translation units. `TILEDB_SERIALIZATION` now
affects what sources files are compiled with ordinary build
dependencies.

The only remaining involvement of `class StorageManager` with `class
RestClient` is an accessor to the `RestClient` instance (currently) held
within its `ContextResources` member variable. Removing this accessor is
out of scope for this PR.

---
TYPE: NO_HISTORY
DESC: `class RestClient` C.41 conformance.

---------

Co-authored-by: KiterLuc <67824247+KiterLuc@users.noreply.github.com>