Remove the superbuild and the external projects. (#5021)
[SC-36912](https://app.shortcut.com/tiledb-inc/story/36912/remove-cmake-variable-tiledb-vcpkg-from-the-build)

[SC-36913](https://app.shortcut.com/tiledb-inc/story/36913/remove-superbuild)

Historically, the Core's build system has been using CMake external
projects to download and build external dependencies, and a "superbuild"
architecture to ensure a build order. With the advent of vcpkg, we have
stopped building the dependencies ourselves and instead rely on vcpkg
(or the "system" in general) to provide them for us. The superbuild has
thus became a relic of the past, consisting of only the inner `tiledb`
project when the new system is enabled (formerly by specifying
`-DTILEDB_VCPKG=ON`, now always).

This PR removes the superbuild. TileDB became a regular CMake project,
whose targets can be built directly without first building the outer
project, and then building the `build/tiledb` subdirectory. The CI
scripts were updated accordingly to not use the subdirectory.

This is inevitably a breaking change in the build system. For starters,
local development environment will certainly need to make a clean build
after this change. Furthermore, there will need to be changes in build
scripts to not build again on the `tiledb` subdirectory.

For examples of downstream migrations, TileDB-Inc/TileDB-Go#316 uses a
`make` invocation that has a similar effect both with and without the
superbuild, and conda-forge/tiledb-feedstock#290 uses a semi-documented
CMake option to disable the superbuild (which will have no effect after
the superbuild gets removed).

The majority of first-party downstreams (VCF, SOMA, MariaDB, Vector
Search, the Python and Java APIs) use the `install-tiledb` target, which
currently is defiend on the superbuild and builds the `install` target
in the inner `tiledb` project. With this PR the `install-tiledb` target
will be kept for compatibility, but alias to `install`.

~~I tried to build VCF with a TileDB external project from this branch,
but it fails with an error that will be fixed with #4989. I will try
again once that PR gets merged.~~ Never mind, VCF builds with the latest
changes.

---
TYPE: BUILD
DESC: The superbuild architecture of the build system has been removed
and TileDB is a regular CMake project. Build commands of the form `make
&& make -C tiledb <targets>` will have to be replaced by `make
<targets>`.