Add specification for delta filter encoding et al. (#5152)
[SC-49777](https://app.shortcut.com/tiledb-inc/story/49777/delta-filter-storage-format-is-undocumented)

This PR adds the encoding of the delta filter to the filter
specification. The specification was based on the filter class'
[implementation](https://github.com/TileDB-Inc/TileDB/blob/6a9918a6caf33967dc3201016ecb51eb62a45d3c/tiledb/sm/compressors/delta_compressor.cc#L225-L245).

The class' documentation had some errors which I fixed. I also fixed the
definition of the delta filter's overhead, and plumbed it to the
`CompressionFilter::overhead` function, and added to the storage format
specification a list of all supported filters.

---
TYPE: NO_HISTORY
DESC: Added specification for the encoding of the delta filter, and a
list of all filters supported by TileDB.