Remove dead code in `StorageManager` and `Query`
With the removal of deprecated function in #5146, the opportunity arises to remove asynchronous query submission, which the C API no longer has any reference to. Alongside this, there is other dead code in `StorageManager` in need of pruning. This PR simplifies its constructor a bit and removes unused members variables.