Remove asynchronous query test. (#5194)
This PR is a follow-on to #4957, which removed tests for the deprecated
function `tiledb_query_submit_async`. That PR removed an entire test
file around that now-removed C API function. In another test in a
different file, though, instead of removing the test, it replaced the
call with the non-async C API function and wrapped it in `std::async`;
this change is not equivalent to the original. This PR does what #4957
did already elsewhere and removes the test.


---
TYPE: NO_HISTORY
DESC: Remove asynchronous query test.