Stop recognizing `sm.use_refactored_readers` config option. (#5183)
[SC-50445](https://app.shortcut.com/tiledb-inc/story/50445/remove-sm-use-refactored-readers-config-option)

This option has been deprecated and undocumented for a while and a
warning was emitted if it is used. This PR removes all mentions of it
from the codebase.

---
TYPE: CONFIG
DESC: The `sm.use_refactored_readers` config option is no longer
recognized. Refactored readers are used by default. To use the legacy
reader, set the
`sm.query_(dense|sparse_global_order|sparse_unordered_with_dups)_reader`
config option (depending on the reader you are using) to `legacy`.