Remove some extraneous capnp dereferences. (#5196)
[sc-48746]

We need to be conservative and follow the CAPNP guidelines on what's
best practice when reading messages: only call `reader.getXYZ` once and
reuse that value so that the traversal counter is not increased and we
don't hit the limit earlier than needed. [Tips and best practices here:
https://capnproto.org/cxx.html]

---
TYPE: NO_HISTORY
DESC: Remove some extraneous capnp dereferences.