Improve dense read performance for small reads. (#5145)
This improves a couples places that are easy gain when doing small dense
reads. Running a small read in a loop (256K tile), I found out a few
places where we can easily cut read times. 2 of those places were nested
parallel fors that have been replaced by a parallel for 2d. The other
was storing a full configuration class, which includes a map of ~148
configuration settings in the subarray. The subarray only uses two
settings and those are now saved as member of the class, removing the
need to generate a full configuration class when a subarray is
constructed.

[sc-50178]

---
TYPE: IMPROVEMENT
DESC: Improve dense read performance for small reads.