Mark Stats::start_timer as [[nodiscard]]. (#5211)
The function `start_timer` returns a sentry object that measures
duration to the end of its enclosing scope. In order to do so, the
object must initialize some variable in the scope. Marking the function
as `[[nodiscard]]` prevents inadvertent misuse of the function. The
documentation was expanded to explain those who might make such an
error.

[sc-51555]

---
TYPE: NO_HISTORY
DESC: Mark Stats::start_timer as [[nodiscard]]