Skip caching redirect uri on array create (#5224)
Currently we store a cache of the region redirection for all requests to
the REST server in `redirect_uri_map`. However, when an array is first
created, we don't know the region yet and we sent the request to the
default region, so we shouldn't be caching this. After the array has
been created, subsequent requests will be redirected by the server to
the right region, and only then the rest client in Core should start
caching in `redirect_uri_map`.

[sc-50254]

---
TYPE: IMPROVEMENT
DESC: Skip caching redirect uri on array create